requirements.txt
Django >= 1.7
furl >= 0.3.95
future
python-social-auth >= 0.1.20
requests >= 2.3.0
